feat(user_proof): Add proof id to UserProof entity
