+++ src/top_app/lib/shared/models/user_specific/user_proof_model.dart
@@ -3,6 +3,7 @@ import 'package:top_app/shared/entities/templates/proof.dart';
 import '../../entities/user_specific/user_proof.dart';
 
 class UserProofModel {
+  final String proofId;
   final ProofType type;
   final String? submittedText;
   final List<String> submittedImageUrls;
@@ -10,6 +11,7 @@ class UserProofModel {
   final bool isValid;
 
   UserProofModel({
+    required this.proofId,
     required this.type,
     this.submittedText,
     required this.submittedImageUrls,
@@ -19,6 +21,7 @@ class UserProofModel {
 
   factory UserProofModel.fromEntity(UserProof entity) {
     return UserProofModel(
+      proofId: entity.proofId,
       type: entity.type,
       submittedText: entity.submittedText,
       submittedImageUrls: entity.submittedImageUrls,
@@ -29,6 +32,7 @@ class UserProofModel {
 
   UserProof toEntity() {
     return UserProof(
+      proofId: proofId,
       type: type,
       submittedText: submittedText,
       submittedImageUrls: submittedImageUrls,
@@ -39,6 +43,7 @@ class UserProofModel {
 
   factory UserProofModel.fromJson(Map<String, dynamic> json) {
     return UserProofModel(
+      proofId: json['proofId'],
       type: ProofType.values.firstWhere(
         (ProofType e) => e.name == json['type'],
         orElse: () => throw Exception('Invalid proof type: ${json['type']}'),
@@ -52,6 +57,7 @@ class UserProofModel {
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{
+      'proofId': proofId,
       'type': type.name,
       'submittedText': submittedText,
       'submittedImageUrls': submittedImageUrls,
